'''Get inspire (http://inspirehep.net) records to fill the bibliography of a LaTeX file
BibTeX is assumed as the default format of the LaTeX file. The bib file name is obtained
from the LaTeX file and the missing BibTeX records from inspire are appended there.

The program requires the \cite key in the "texkey" format of inspire (Lastname:yearcode).
However if a \cite has an eprint number as key like hep-ph/0213456 or 1206:0001, it will
be automatically converted to a "texkey" and both the LaTeX file and the bib file will
be updated.

The BibTeX bbl generated file can be changed by one with The LaTeX records inspire
format (see --bbl option)

The \\begin{thebibliography} format can be also processed with the output in an external
bbl file (see -t option)